requirements.txt
streamlit>=1.32
yfinance>=0.2.36
plotly>=5.18
pandas>=2.0
nltk>=3.8
praw>=7.7
python-dotenv>=1.0
requests>=2.31
finnhub-python>=2.4
flask>=3.0
gunicorn>=21.0
apscheduler>=3.10
